+++ pcs/lib/env.py
@@ -1,24 +1,11 @@
 from logging import Logger
-from typing import (
-    Any,
-    Callable,
-    Mapping,
-    Optional,
-    Union,
-    cast,
-)
+from typing import Any, Callable, Mapping, Optional, Union, cast
 
 from lxml.etree import _Element
 
-from pcs.common import (
-    file_type_codes,
-    reports,
-)
+from pcs.common import file_type_codes, reports
 from pcs.common.host import PcsKnownHost
-from pcs.common.node_communicator import (
-    Communicator,
-    NodeCommunicatorFactory,
-)
+from pcs.common.node_communicator import Communicator, NodeCommunicatorFactory
 from pcs.common.reports import ReportProcessor
 from pcs.common.reports.item import ReportItem
 from pcs.common.services.interfaces import ServiceManagerInterface
@@ -31,10 +18,7 @@ from pcs.lib.communication.corosync import (
     DistributeCorosyncConf,
     ReloadCorosyncConf,
 )
-from pcs.lib.communication.tools import (
-    run,
-    run_and_raise,
-)
+from pcs.lib.communication.tools import run, run_and_raise
 from pcs.lib.corosync.config_facade import ConfigFacade as CorosyncConfigFacade
 from pcs.lib.corosync.config_parser import (
     verify_section as verify_corosync_section,
@@ -245,15 +229,21 @@ class LibraryEnvironment:
                 ReportItem.error(reports.messages.WaitForIdleNotLiveCluster())
             )
 
-    def push_cib(self, custom_cib=None, wait_timeout: int = -1) -> None:
+    def push_cib(
+        self,
+        custom_cib: Optional[_Element] = None,
+        wait_timeout: int = -1,
+        with_status: bool = False,
+    ) -> None:
         """
         Push previously loaded instance of CIB or a custom CIB
 
-        etree custom_cib -- push a custom CIB instead of a loaded instance
+        custom_cib -- push a custom CIB instead of a loaded instance
             (allows to push an externally provided CIB and replace the one in
             the cluster completely)
         wait_timeout -- wait timeout in seconds, if less than 0 wait will be
             skipped, if 0 wait indefinitely
+        with_status -- push also status section of a CIB
         """
         self._ensure_wait_satisfiable(wait_timeout)
         if custom_cib is not None:
@@ -261,10 +251,14 @@ class LibraryEnvironment:
                 raise AssertionError(
                     "CIB has been loaded, cannot push custom CIB"
                 )
+            if with_status:
+                raise AssertionError(
+                    "Cannot push status section of a custom CIB"
+                )
             return self.__push_cib_full(custom_cib, wait_timeout)
         if self.__loaded_cib_diff_source is None:
             raise AssertionError("CIB has not been loaded")
-        return self.__push_cib_diff(wait_timeout)
+        return self.__push_cib_diff(wait_timeout, with_status)
 
     def __push_cib_full(self, cib_to_push, wait_timeout: int):
         self.__do_push_cib(
@@ -272,20 +266,23 @@ class LibraryEnvironment:
             wait_timeout,
         )
 
-    def __push_cib_diff(self, wait_timeout: int):
+    def __push_cib_diff(self, wait_timeout: int, with_status: bool = False):
         self.__do_push_cib(
-            lambda: self.__main_push_cib_diff(self.cmd_runner()), wait_timeout
+            lambda: self.__main_push_cib_diff(self.cmd_runner(), with_status),
+            wait_timeout,
         )
 
-    def __main_push_cib_diff(self, cmd_runner):
+    def __main_push_cib_diff(
+        self, cmd_runner: CommandRunner, with_status: bool = False
+    ):
         cib_diff_xml = diff_cibs_xml(
             cmd_runner,
             self.report_processor,
-            self.__loaded_cib_diff_source,
-            etree_to_str(self.__loaded_cib_to_modify),
+            cast(str, self.__loaded_cib_diff_source),
+            etree_to_str(cast(_Element, self.__loaded_cib_to_modify)),
         )
         if cib_diff_xml:
-            push_cib_diff_xml(cmd_runner, cib_diff_xml)
+            push_cib_diff_xml(cmd_runner, cib_diff_xml, with_status)
 
     def __do_push_cib(self, push_strategy, wait_timeout: int) -> None:
         push_strategy()